refactor: TPO 코디 추천 트랜잭션 분리 및 예외 처리 개선   - TpoSaveService 추출: DB 저장을 AI-BE 호출 이후 별도 트랜잭션으로 처리   - generateTpoOutfits에 NOT_SUPPORTED 적용, DB 커넥션 장시간 점유 방지   - HttpAIClient.recommendOutfit 예외 처리를 searchShop과 동일하게 통일
